requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
requests==2.32.3
google-api-python-client==2.145.0
google-auth==2.34.0
